+++ src/js/lib/card-nicknames/neo.ts
@@ -1,0 +1,155 @@
+// Kamigawa: Neon Dynasty (NEO)
+export default [
+  {
+    // https://scryfall.com/card/neo/35/ssselfless-ssshmamurai?utm_source=api
+    setCode: "neo",
+    collectorNumber: "35",
+    realName: ["Selfless Samurai"],
+    nickname: ["Ssselfless Ssshmamurai"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/neo/37/sleepy-moth-fly-guy?utm_source=api
+    setCode: "neo",
+    collectorNumber: "37",
+    realName: ["Sky-Blessed Samurai"],
+    nickname: ["Sleepy Moth Fly Guy"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/neo/70/master-of-doritos?utm_source=api
+    setCode: "neo",
+    collectorNumber: "70",
+    realName: ["Moonsnare Specialist"],
+    nickname: ["Master of Doritos"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/neo/122/soul-slurpee%C2%AE?utm_source=api
+    setCode: "neo",
+    collectorNumber: "122",
+    realName: ["Soul Transfer"],
+    nickname: ["Soul Slurpee®"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/neo/140/praise-the-sun?utm_source=api
+    setCode: "neo",
+    collectorNumber: "140",
+    realName: ["Explosive Singularity"],
+    nickname: ["Praise the Sun"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/neo/155/crab-ogre-helmet?utm_source=api
+    setCode: "neo",
+    collectorNumber: "155",
+    realName: ["Ogre-Head Helm"],
+    nickname: ["Crab-Ogre Helmet"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/neo/157/dharma-rabbit-%2315?utm_source=api
+    setCode: "neo",
+    collectorNumber: "157",
+    realName: ["Rabbit Battery"],
+    nickname: ["Dharma Rabbit #15"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/neo/159/goblin-manicurist?utm_source=api
+    setCode: "neo",
+    collectorNumber: "159",
+    realName: ["Scrap Welder"],
+    nickname: ["Goblin Manicurist"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/neo/161/explosive-diarrhea?utm_source=api
+    setCode: "neo",
+    collectorNumber: "161",
+    realName: ["Seismic Wave"],
+    nickname: ["Explosive Diarrhea"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/neo/166/flower-delivery-raiju?utm_source=api
+    setCode: "neo",
+    collectorNumber: "166",
+    realName: ["Thundering Raiju"],
+    nickname: ["Flower-Delivery Raiju"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/neo/175/bubblegum-stag?utm_source=api
+    setCode: "neo",
+    collectorNumber: "175",
+    realName: ["Blossom Prancer"],
+    nickname: ["Bubblegum Stag"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/neo/203/quadruple-orochi?utm_source=api
+    setCode: "neo",
+    collectorNumber: "203",
+    realName: ["Orochi Merge-Keeper"],
+    nickname: ["Quadruple Orochi"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/neo/214/weaver-bracelet?utm_source=api
+    setCode: "neo",
+    collectorNumber: "214",
+    realName: ["Webspinner Cuff"],
+    nickname: ["Weaver Bracelet"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/neo/220/shiga-big-deal-biker?utm_source=api
+    setCode: "neo",
+    collectorNumber: "220",
+    realName: ["Greasefang, Okiba Boss"],
+    nickname: ["Shiga, Big Deal Biker"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/neo/223/good-soup?utm_source=api
+    setCode: "neo",
+    collectorNumber: "223",
+    realName: ["Invigorating Hot Spring"],
+    nickname: ["Good Soup"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/neo/240/seems-an-awful-waste?utm_source=api
+    setCode: "neo",
+    collectorNumber: "240",
+    realName: ["Bronze Cudgels"],
+    nickname: ["Seems An Awful Waste..."],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/neo/254/reassembling-automation?utm_source=api
+    setCode: "neo",
+    collectorNumber: "254",
+    realName: ["Patchwork Automaton"],
+    nickname: ["Reassembling Automation"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/neo/263/leggy-temple?utm_source=api
+    setCode: "neo",
+    collectorNumber: "263",
+    realName: ["Walking Skyscraper"],
+    nickname: ["Leggy Temple"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/neo/276/sokenzan-sizzling-barbecue?utm_source=api
+    setCode: "neo",
+    collectorNumber: "276",
+    realName: ["Sokenzan, Crucible of Defiance"],
+    nickname: ["Sokenzan, Sizzling Barbecue"],
+    source: "Scryfall Preview Name",
+  },
+];